// packages/babylon-proto-ts/src/generated/babylon/epoching/v1/tx.ts
export interface MsgWrappedDelegate {
  msg: MsgDelegate | undefined;
}

/** MsgWrappedDelegate is the response to the MsgWrappedDelegate message */
export interface MsgWrappedDelegateResponse {
}

/** MsgWrappedUndelegate is the message for undelegating stakes */
export interface MsgWrappedUndelegate {
  msg: MsgUndelegate | undefined;
}
// ...
export const MsgWrappedDelegate: MessageFns<MsgWrappedDelegate> = {
  encode(message: MsgWrappedDelegate, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.msg !== undefined) {
      MsgDelegate.encode(message.msg, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgWrappedDelegate {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgWrappedDelegate();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.msg = MsgDelegate.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgWrappedDelegate {
    return { msg: isSet(object.msg) ? MsgDelegate.fromJSON(object.msg) : undefined };
  },

  toJSON(message: MsgWrappedDelegate): unknown {
    const obj: any = {};
    if (message.msg !== undefined) {
      obj.msg = MsgDelegate.toJSON(message.msg);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgWrappedDelegate>, I>>(base?: I): MsgWrappedDelegate {
    return MsgWrappedDelegate.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgWrappedDelegate>, I>>(object: I): MsgWrappedDelegate {
    const message = createBaseMsgWrappedDelegate();
    message.msg = (object.msg !== undefined && object.msg !== null) ? MsgDelegate.fromPartial(object.msg) : undefined;
    return message;
  },
};
// ...
export const MsgWrappedUndelegate: MessageFns<MsgWrappedUndelegate> = {
  encode(message: MsgWrappedUndelegate, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.msg !== undefined) {
      MsgUndelegate.encode(message.msg, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgWrappedUndelegate {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgWrappedUndelegate();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.msg = MsgUndelegate.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgWrappedUndelegate {
    return { msg: isSet(object.msg) ? MsgUndelegate.fromJSON(object.msg) : undefined };
  },

  toJSON(message: MsgWrappedUndelegate): unknown {
    const obj: any = {};
    if (message.msg !== undefined) {
      obj.msg = MsgUndelegate.toJSON(message.msg);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgWrappedUndelegate>, I>>(base?: I): MsgWrappedUndelegate {
    return MsgWrappedUndelegate.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgWrappedUndelegate>, I>>(object: I): MsgWrappedUndelegate {
    const message = createBaseMsgWrappedUndelegate();
    message.msg = (object.msg !== undefined && object.msg !== null) ? MsgUndelegate.fromPartial(object.msg) : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}